backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.2
neo4j==5.14.0
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
pytesseract==0.3.10
